Add JSON utility functions, Viper configuration reader, IP utility functions, and file utility functions
