+++ pkg/utils/conf/viper.go
@@ -1,4 +1,4 @@
-package utils
+package conf
 
 import (
 	"github.com/spf13/viper"
