+++ service/singleton/notification.go
@@ -1,8 +1,8 @@
 package singleton
 
 import (
-	nn "go-gin/pkg/notification"
-	"go-gin/pkg/utils"
+	nfn "go-gin/pkg/notification"
+	"go-gin/pkg/utils/parse"
 )
 
 var Notifications map[string][]interface{}
@@ -14,27 +14,27 @@ func LoadNotifications() {
 		switch _notification.Type {
 		case "apprise":
 			for _, instance := range _notification.Instances {
-				Notifications["apprise"] = append(Notifications["apprise"], nn.Apprise{Payload: nn.ApprisePayload{AppriseUrl: instance["url"]}})
+				Notifications["apprise"] = append(Notifications["apprise"], nfn.Apprise{Payload: nfn.ApprisePayload{AppriseUrl: instance["url"]}})
 			}
 		case "dingtalk":
 			for _, instance := range _notification.Instances {
-				Notifications["dingtalk"] = append(Notifications["dingtalk"], nn.DingTalk{Payload: nn.DingTalkPayload{Webhook: instance["webhook"]}})
+				Notifications["dingtalk"] = append(Notifications["dingtalk"], nfn.DingTalk{Payload: nfn.DingTalkPayload{Webhook: instance["webhook"]}})
 			}
 		case "ifttt":
 			for _, instance := range _notification.Instances {
-				Notifications["ifttt"] = append(Notifications["ifttt"], nn.IFTTT{Payload: nn.IFTTTPayload{Event: instance["event"], Key: instance["key"]}})
+				Notifications["ifttt"] = append(Notifications["ifttt"], nfn.IFTTT{Payload: nfn.IFTTTPayload{Event: instance["event"], Key: instance["key"]}})
 			}
 		case "telegram":
 			for _, instance := range _notification.Instances {
-				Notifications["telegram"] = append(Notifications["telegram"], nn.Telegram{Payload: nn.TelegramPayload{BotToken: instance["token"], ChatId: instance["chat_id"]}})
+				Notifications["telegram"] = append(Notifications["telegram"], nfn.Telegram{Payload: nfn.TelegramPayload{BotToken: instance["token"], ChatId: instance["chat_id"]}})
 			}
 		case "wecom":
 			for _, instance := range _notification.Instances {
-				Notifications["wecom"] = append(Notifications["wecom"], nn.WeCom{Payload: nn.WeComPayload{Key: instance["key"]}})
+				Notifications["wecom"] = append(Notifications["wecom"], nfn.WeCom{Payload: nfn.WeComPayload{Key: instance["key"]}})
 			}
 		case "smtp":
 			for _, instance := range _notification.Instances {
-				Notifications["smtp"] = append(Notifications["smtp"], nn.SMTP{Payload: nn.SMTPPayload{Host: instance["host"], Port: utils.ParseInt(instance["port"], 587), Username: instance["username"], Password: instance["password"], From: instance["from"], To: instance["to"]}})
+				Notifications["smtp"] = append(Notifications["smtp"], nfn.SMTP{Payload: nfn.SMTPPayload{Host: instance["host"], Port: parse.ParseInt(instance["port"], 587), Username: instance["username"], Password: instance["password"], From: instance["from"], To: instance["to"]}})
 			}
 		default:
 			Log.Error().Msgf("Unknown notification type: %s", _notification.Type)
@@ -47,17 +47,17 @@ func SendNotification(title string, message string) {
 	for _, _notification := range Notifications {
 		for _, instance := range _notification {
 			switch _instance := instance.(type) {
-			case nn.Apprise:
+			case nfn.Apprise:
 				_instance.Send(title, message)
-			case nn.DingTalk:
+			case nfn.DingTalk:
 				_instance.Send(title, message)
-			case nn.IFTTT:
+			case nfn.IFTTT:
 				_instance.Send(title, message)
-			case nn.Telegram:
+			case nfn.Telegram:
 				_instance.Send(title, message)
-			case nn.WeCom:
+			case nfn.WeCom:
 				_instance.Send(title, message)
-			case nn.SMTP:
+			case nfn.SMTP:
 				_instance.Send(title, message)
 			default:
 				Log.Error().Msgf("Unknown notification instance: %v", _instance)
@@ -70,17 +70,17 @@ func SendNotification(title string, message string) {
 func SendNotificationByType(notificationType string, title string, message string) {
 	for _, instance := range Notifications[notificationType] {
 		switch _instance := instance.(type) {
-		case nn.Apprise:
+		case nfn.Apprise:
 			_instance.Send(title, message)
-		case nn.DingTalk:
+		case nfn.DingTalk:
 			_instance.Send(title, message)
-		case nn.IFTTT:
+		case nfn.IFTTT:
 			_instance.Send(title, message)
-		case nn.Telegram:
+		case nfn.Telegram:
 			_instance.Send(title, message)
-		case nn.WeCom:
+		case nfn.WeCom:
 			_instance.Send(title, message)
-		case nn.SMTP:
+		case nfn.SMTP:
 			_instance.Send(title, message)
 		default:
 			Log.Error().Msgf("Unknown notification instance: %v", _instance)
